Refactor and improve sadu-queries2 tests
The commit includes a significant refactoring of the `ParsedExample`, `ParameterExample` and `QueryExample` classes in the sadu-queries2 tests suite, improving their interactions with the database. Specific changes include using the `PostgresDatabase` class for establishing connections, and using the `Reader` class for executing queries and mapping results. Moreover, some tests were disabled, awaiting further modifications.
