+++ sadu-queries2/src/test/java/de/chojo/sadu/queries/examples/ParsedExample.java
@@ -6,83 +6,61 @@
 
 package de.chojo.sadu.queries.examples;
 
-import de.chojo.sadu.databases.PostgreSql;
-import de.chojo.sadu.datasource.DataSourceCreator;
+import de.chojo.sadu.PostgresDatabase;
 import de.chojo.sadu.mapper.PostgresqlMapper;
 import de.chojo.sadu.mapper.RowMapperRegistry;
-import de.chojo.sadu.mapper.rowmapper.RowMapper;
+import de.chojo.sadu.queries.api.execution.reading.Reader;
+import de.chojo.sadu.queries.api.results.writing.ManipulationBatchResult;
 import de.chojo.sadu.queries.call.Call;
 import de.chojo.sadu.queries.calls.Calls;
 import de.chojo.sadu.queries.configuration.QueryConfiguration;
 import de.chojo.sadu.queries.configuration.QueryConfigurationBuilder;
-import de.chojo.sadu.queries.stages.execution.reading.QueryReader;
-import de.chojo.sadu.queries.stages.results.writing.ManipulationBatchQuery;
+import de.chojo.sadu.queries.examples.dao.User;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
-import org.testcontainers.containers.GenericContainer;
 
-import javax.sql.DataSource;
 import java.io.IOException;
-import java.util.List;
+import java.sql.SQLException;
 import java.util.UUID;
 
-import static de.chojo.sadu.queries.PostgresDatabase.createContainer;
+import static de.chojo.sadu.PostgresDatabase.createContainer;
 import static de.chojo.sadu.queries.call.adapter.impl.UUIDAdapter.AS_BYTES;
 
 public class ParsedExample {
-    private static GenericContainer<?> pg;
     private static QueryConfiguration query;
+    private static PostgresDatabase.Database db;
 
     @BeforeAll
-    static void beforeAll() throws IOException {
-        pg = createContainer("postgres", "postgres");
-        DataSource dc = DataSourceCreator.create(PostgreSql.get())
-                .configure(c -> c.host(pg.getHost()).port(pg.getFirstMappedPort())).create()
-                .usingPassword("postgres")
-                .usingUsername("postgres")
-                .build();
-
-        List<RowMapper<?>> defaultMapper = PostgresqlMapper.getDefaultMapper();
-        RowMapperRegistry registry = new RowMapperRegistry().register(defaultMapper);
-        query = new QueryConfigurationBuilder(dc).setRowMapperRegistry(registry).build();
+    static void beforeAll() throws IOException, SQLException {
+        db = createContainer("postgres", "postgres");
+        query = new QueryConfigurationBuilder(db.dataSource()).setRowMapperRegistry(new RowMapperRegistry().register(PostgresqlMapper.getDefaultMapper())).build();
     }
 
     @AfterAll
     static void afterAll() throws IOException {
-        pg.close();
+        db.close();
     }
 
     @Test
+    @Disabled
     public void example() {
         // Executing a single call by directly creating the call
-        QueryReader<User> parameter = query.query("SELECT * FROM table WHERE id = ?, name ILIKE :name)")
+        Reader<User> parameter = query.query("SELECT * FROM table WHERE id = ?, name ILIKE :name)")
                 .single(Calls.single(c -> c.bind("some input").bind("name", "user")))
-                .map(row -> new User(row.getInt("id"), row.getUuidFromBytes("uuid"), row.getString("name")));
+                .map(User::map);
 
         // Map the result to some user object
-        QueryReader<User> parameter1 = query.query("SELECT * FROM table where uuid = :uuid")
+        Reader<User> parameter1 = query.query("SELECT * FROM table where uuid = :uuid")
                 .single(Calls.single(c -> c.bind("uuid", UUID.randomUUID(), AS_BYTES)))
-                .map(row -> new User(row.getInt("id"), row.getUuidFromBytes("uuid"), row.getString("name")));
+                .map(User::map);
 
         // Perform an insert
-        ManipulationBatchQuery insert = query.query("INSERT INTO table VALUES(?)")
+        ManipulationBatchResult insert = query.query("INSERT INTO table VALUES(?)")
                 .batch(Calls.batch(
                         Call.of().bind(UUID.randomUUID(),AS_BYTES),
                         Call.of().bind(UUID.randomUUID(),AS_BYTES)))
                 .insert();
-    }
-
-
-    public class User{
-        private final int id;
-        private final UUID uuid;
-        private final String name;
-
-        public User(int id, UUID uuid, String name) {
-            this.id = id;
-            this.uuid = uuid;
-            this.name = name;
-        }
     }
 }
