+++ sadu-queries2/src/test/java/de/chojo/sadu/queries/examples/QueryExample.java
@@ -6,48 +6,38 @@
 
 package de.chojo.sadu.queries.examples;
 
-import de.chojo.sadu.databases.PostgreSql;
-import de.chojo.sadu.datasource.DataSourceCreator;
+import de.chojo.sadu.PostgresDatabase;
 import de.chojo.sadu.mapper.PostgresqlMapper;
 import de.chojo.sadu.mapper.RowMapperRegistry;
-import de.chojo.sadu.mapper.rowmapper.RowMapper;
 import de.chojo.sadu.queries.configuration.QueryConfiguration;
 import de.chojo.sadu.queries.configuration.QueryConfigurationBuilder;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
-import org.testcontainers.containers.GenericContainer;
 
-import javax.sql.DataSource;
 import java.io.IOException;
-import java.util.List;
+import java.sql.SQLException;
 
-import static de.chojo.sadu.queries.PostgresDatabase.createContainer;
+import static de.chojo.sadu.PostgresDatabase.createContainer;
 
 class QueryExample {
-    private static GenericContainer<?> pg;
     private static QueryConfiguration query;
+    private static PostgresDatabase.Database db;
 
     @BeforeAll
-    static void beforeAll() throws IOException {
-        pg = createContainer("postgres", "postgres");
-        DataSource dc = DataSourceCreator.create(PostgreSql.get())
-                .configure(c -> c.host(pg.getHost()).port(pg.getFirstMappedPort())).create()
-                .usingPassword("postgres")
-                .usingUsername("postgres")
-                .build();
-
-        List<RowMapper<?>> defaultMapper = PostgresqlMapper.getDefaultMapper();
-        RowMapperRegistry registry = new RowMapperRegistry().register(defaultMapper);
-        query = new QueryConfigurationBuilder(dc).setRowMapperRegistry(registry).build();
+    static void beforeAll() throws IOException, SQLException {
+        db = createContainer("postgres", "postgres");
+        query = new QueryConfigurationBuilder(db.dataSource()).setRowMapperRegistry(new RowMapperRegistry().register(PostgresqlMapper.getDefaultMapper())).build();
     }
 
     @AfterAll
     static void afterAll() throws IOException {
-        pg.close();
+        db.close();
     }
 
     @Test
+    @Disabled
     public void example() {
         // A query with a named parameter
         query.query("SELECT * FROM table where uuid = :uuid");
